+++ app.py
@@ -11,77 +11,15 @@ from chalicelib.controllers.sales_crontoller import (
     list_sale, create_sale, filter_sale_by_id, update_sale,
     delete_sale, get_sales_by_products, get_all_sales_by_products
 )
+from chalicelib.routes import unit_measure_route
 
 
 app = Chalice(app_name='challenge')
+app.register_blueprint(unit_measure_route.unit_measure_routes)
 
 @app.route('/')
 def index():
     return {'Welcome': 'API Challenge'}
-
-
-@app.route('/unit-measure', methods=['GET'])
-def get_all_unit_measure():
-    get_unit_measure, status_code = list_unit_measure()
-    return Response(
-        body=get_unit_measure,
-        status_code=status_code,
-    )
-
-
-@app.route('/unit-measure', methods=['POST'])
-def add_unit_measure():
-    data = app.current_request.json_body
-    name = data.get('name')
-    if not (name):
-        return Response(
-            body={
-                "message": "Name factor are required",
-                "error": "Incorrect request body."
-            },
-            status_code=400,
-        )
-
-    response_data, status_code = create_unit_measure(name=name)
-    return Response(
-        body=response_data,
-        status_code=status_code,
-    )
-
-
-@app.route('/unit-measure/{id}', methods=['GET'])
-def get_unit_measure_by_id(id):
-    get_unit_measure, status_code = filter_unit_measure_by_id(id=id)
-    return Response(
-        body=get_unit_measure,
-        status_code=status_code,
-    )
-
-
-@app.route('/unit-measure/{id}', methods=['PUT'])
-def update_unit_measure_by_id(id):
-    data = app.current_request.json_body
-    name = data.get('name')
-    if not (name):
-        return Response(
-            body={'error': 'Name factor are required.'},
-            status_code=400,
-        )
-
-    response_data, status_code = update_unit_measure(id=id, name=name)
-    return Response(
-        body=response_data,
-        status_code=status_code,
-    )
-
-
-@app.route('/unit-measure/{id}', methods=['DELETE'])
-def delete_unit_measure_by_id(id):
-    response_data, status_code = delete_unit_measure(id=id)
-    return Response(
-        body=response_data,
-        status_code=status_code,
-    )
 
 
 # endpoint for product
